changed type of loop from int to size_t to ensure type matching

+++ src/Inciter/Partitioner.C
@@ -871,7 +871,7 @@ Partitioner::prepare()
         // TODO: We can replace this linear search with a float hash map
         // Search for matching coords
         bool not_found = true;
-        for (int i = 0; i < coord_x.size(); i++)
+        for (size_t i = 0; i < coord_x.size(); i++)
         {
             tk::real x = coord_x[i];
             tk::real y = coord_y[i];
